+++ src/main/java/br/com/letscode/dbletscode/controller/DisciplinaRestController.java
@@ -8,10 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,9 +29,9 @@ public class DisciplinaRestController {
         return disciplinaRepository.findAll();
     }
 
-    @GetMapping("/findbyid")
+    @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.FOUND)
-    public Optional<Disciplina> findById(@RequestParam String id) {
+    public Optional<Disciplina> findById(@PathVariable String id) {
         return disciplinaRepository.findById(id);
     }
 
@@ -41,17 +41,18 @@ public class DisciplinaRestController {
         return disciplinaRepository.save(disciplina);
     }
 
-    @DeleteMapping
-    public void deleteById(@RequestBody String id) {
-        disciplinaRepository.deleteById(id);
+    @DeleteMapping("/{id}")
+    public void deleteById(@PathVariable String id) {
+        if (isExists(id)) disciplinaRepository.deleteById(id);
     }
 
-    @PatchMapping
-    public Disciplina updateDisciplina(@RequestBody Disciplina disciplina) {
-        if (disciplinaRepository.existsById(disciplina.getCodigoDisciplina())) {
-            this.deleteById(disciplina.getCodigoDisciplina());
-            return this.inserirDisciplina(disciplina);
-        }
+    @PatchMapping("/{id}")
+    public Disciplina updateDisciplina(@PathVariable String id, @RequestBody Disciplina disciplina) {
+        if (isExists(id)) return this.inserirDisciplina(disciplina);
         else throw new IAmATeapot();
+    }
+
+    private boolean isExists(String id) {
+        return disciplinaRepository.existsById(id);
     }
 }
