+++ src/main/java/br/com/letscode/dbletscode/controller/ProfessorRestController.java
@@ -8,10 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,9 +29,9 @@ public class ProfessorRestController {
         return professorRepository.findAll();
     }
 
-    @GetMapping("/findbyid")
+    @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.FOUND)
-    public Optional<Professor> findById(@RequestParam Integer id) {
+    public Optional<Professor> findById(@PathVariable int id) {
         return professorRepository.findById(id);
     }
 
@@ -41,17 +41,18 @@ public class ProfessorRestController {
         return professorRepository.save(professor);
     }
 
-    @DeleteMapping
-    public void deleteById(@RequestBody Integer id) {
-        professorRepository.deleteById(id);
+    @DeleteMapping("/{id}")
+    public void deleteById(@PathVariable int id) {
+        if (isExists(id)) professorRepository.deleteById(id);
     }
 
-    @PatchMapping
-    public Professor updateProfessor(@RequestBody Professor professor) {
-        if (professorRepository.existsById(professor.getRegistroProfessor())) {
-            this.deleteById(professor.getRegistroProfessor());
-            return this.inserirProfessor(professor);
-        }
+    @PatchMapping("/{id}")
+    public Professor updateProfessor(@PathVariable int id, @RequestBody Professor professor) {
+        if (isExists(id)) return this.inserirProfessor(professor);
         else throw new IAmATeapot();
+    }
+
+    private boolean isExists(int id) {
+        return professorRepository.existsById(id);
     }
 }
